Task3-Housing-Price-Multimodal/requirements.txt
torch==2.0.0
torchvision==0.15.0
gradio==4.31.0
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
matplotlib==3.8.4
Pillow==10.3.0
joblib==1.4.0
